Local development improvements
+ add a start-firefox command. uses the selenium web driver to start
firefox with a good profile.
+ add a "not connected" pane for telling people to connect firefox.
+ add openurl, for automatically opening the debugger when the local
server is started
+ cleaned up the readme getting started. Now just 3 commands!
+ added testing / linting commands to the readme.
