+++ README.md
@@ -6,26 +6,19 @@ This is a prototype debugger written without any XUL and based on React and Redu
 
 #### Getting Started
 
-```
-$ npm install
-$ npm start
-```
+* `npm install` - Install Dependencies
+* `npm run start-firefox` - Start Firefox
+* `npm start` - Start Debugger
 
-Start Firefox in remote debugging mode. The `-P` parameter specifies a profile to use:
+![screen shot 2016-05-16 at 1 24 29 pm](https://cloud.githubusercontent.com/assets/254562/15297643/34575ca6-1b69-11e6-9703-8ba0a029d4f9.png)
 
-```
-$ /Applications/FirefoxNightly.app/Contents/MacOS/firefox-bin -P development --start-debugger-server 6080
-```
-
-Set additional configuration options in Firefox using `about:config`:
-
-- `devtools.debugger.remote-enabled` to `true`
-- `devtools.chrome.enabled` to `true`
-- `devtools.debugger.prompt-connection` to `false`
-
-Quit and re-open firefox with the same command. Go to some pages like [todomvc](http://todomvc.com/examples/backbone/) or [nyt](http://www.nytimes.com/).
-
-Then open another browser and go to `http://localhost:8000`.
+#### Running tests
+* `npm test` - Run unit tests
+* `npm run mocha-server` - Run unit tests in the browser
+* `cypress run` - Run integration tests
+* `cypress open` - Run integration tests in the browser
+* `npm run lint` - Run CSS and JS linter
+* `npm run storybook` - Open Storybook
 
 #### Advanced :see_no_evil:
 
@@ -37,3 +30,25 @@ You can create a `development.local.json` for local user settings in `public/js/
 
 ##### Remote Debugging
 If you'd like to connect an existing Firefox browser to debugger.html, you can press `shift+F2` to open the developer toolbar and type `listen 6080` into the developer toolbar console.
+
+##### Starting Firefox
+
+Sometimes you will want to open firefox manually.
+
+1) open a specific version of firefox
+2) use a different profile
+
+It is easy to open firefox with the `firefox-bin` script:
+
+```
+$ /Applications/FirefoxNightly.app/Contents/MacOS/firefox-bin -P development --start-debugger-server 6080
+```
+
+* *--start-debugger-server 6080* Start Firefox in remote debugging mode.
+* *-P development* parameter specifies a profile to use:
+
+Firefox needs to some settings configured in `about:config` to remotely connect to devtools:
+
+- `devtools.debugger.remote-enabled` to `true`
+- `devtools.chrome.enabled` to `true`
+- `devtools.debugger.prompt-connection` to `false`
