+++ index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
+    <title>Firefox Debugger</title>
     <meta charset="utf8" />
     <link rel="stylesheet" href="public/build/styles.css" type="text/css" />
   </head>
