+++ public/js/components/App.css
@@ -17,6 +17,14 @@ body {
   width: 100%;
 }
 
+.not-connected-message {
+  margin: auto;
+  padding: 50px 100px;
+  border: 1px solid #dddddd;
+  background-color: #fbfbfb;
+  color: #9a9a9a;
+}
+
 #mount {
   display: flex;
   height: 100%;
